+++ creator-node/src/components/healthCheck/healthCheckComponentService.js
@@ -1,6 +1,12 @@
 const versionInfo = require('../../../.version.json')
 const { getLogger } = require('../../logging.js')
 
+/**
+ * Perform a basic health check, returning the
+ * currently selected discovery provider (if any),
+ * the current git SHA, and service version info.
+ * @param {*} ServiceRegistry
+ */
 const healthCheck = ({ libs } = {}) => {
   const logger = getLogger()
